refactor(db): split disk serialization types into their own module
